backend/requirements.txt
async-timeout==4.0.2
attrs==23.1.0
exceptiongroup==1.1.2
greenlet==2.0.2
iniconfig==2.0.0
logzero==1.7.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pyscopg2==66.0.2
pytest==7.4.0
python-decouple==3.8
redis==4.6.0
referencing==0.30.0
rpds-py==0.9.2
secure-smtplib==0.1.1
SQLAlchemy==2.0.19
tomli==2.0.1
tornado==6.3.2
tornado-cors==0.6.0
torndb==0.3
typing_extensions==4.7.1
